requirements.txt
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
biopython==1.83
bleach==6.1.0
bokeh==3.4.1
cassiopeia-lineage @ git+https://github.com/YosefLab/Cassiopeia@41dbff8886c886dd1972347f05e29fe8b9a3f0fc
certifi==2024.2.2
charset-normalizer==3.3.2
clarabel==0.7.1
comm==0.2.2
contourpy==1.2.1
cvxpy==1.4.3
cycler==0.12.1
Cython==3.0.10
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
ecos==2.0.13
ete3==3.1.3
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.19.1
fonttools==4.51.0
hits==0.4.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.18.1
ipywidgets==8.1.2
itolapi==4.1.4
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
llvmlite==0.41.1
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
ngs-tools==1.8.5
numba==0.58.1
numpy==1.26.4
osqp==0.6.5
packaging @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_a6lqg7at4g/croot/packaging_1710807410750/work
pandas==2.2.2
pandocfilters==1.5.1
parameterized==0.9.0
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.1
plotly==5.21.0
ply==3.11
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pybind11==2.12.0
Pygments==2.17.2
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-sip @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_7c_9f71up5/croot/pyqt-split_1698769091879/work/pyqt_sip
pysam==0.22.1
pyseq-align==1.0.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.2
qdldl==0.1.7.post2
referencing==0.35.0
requests==2.31.0
rpds-py==0.18.0
scipy==1.11.4
scs==3.2.4.post1
seaborn==0.13.2
shortuuid==1.0.13
sip @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_ba9odc_3u1/croot/sip_1698675938651/work
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
statsmodels==0.14.2
tenacity==8.2.3
tinycss2==1.3.0
tomli @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_d0e5ffbf-5cf1-45be-8693-c5dff8108a2awhthtjlq/croots/recipe/tomli_1657175508477/work
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
widgetsnbextension==4.0.10
xyzservices==2024.4.0
zipp==3.18.1
